requirements.txt
django==1.5.1
django-social-auth
